backend/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0
wikipedia>=1.4.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
